Support multiple chef servers in a aws accound in the same region.
Basically we need to be able to ignore messages that bootstrap against different chef servers in the same aws account and region. We accomplish this with the use of aws tags. This is a breaking change in the sense now you must specify a tag. Here is a couple examples:
- `./bin/aws_watcher.rb -t 'environment:dev,qa'`
- `./bin/aws_watcher.rb -t 'environment:prod-west-1,prod-east-1'`
- `./bin/aws_watcher.rb -t 'chef_server:my_chef_server_name'`